Slo sammen ReglerForMedl og ReglerForRegistrerteOpplysninger.kt
